+++ packages/ragbits-chat/CHANGELOG.md
@@ -3,6 +3,14 @@
 ## Unreleased
 - Split params into path params, query params in API client (#871)
 - Fix bug causing conversation not to be selected when navigating to it from url(#872)
+- CI/CD for nightly npm builds, update ragbits-api-client-react deps to install latest version of the @ragbits/api-client (#873)
+- CI/CD for nightlty builds improvements (#874)
+- Add automatic topic extraction to be used as conversation title with ability to edit in the client side (#840)
+- Add todo list component to the UI, add support for todo events in API (#827)
+- Add support for confirmation requests in chat (#853)
+- customizable HeroUI theme (#841)
+- Add error response type to the chat interface with ability to display error messages to the user (#878)
+
 ### Added
 
 - Add automatic topic extraction to be used as conversation title with ability to edit in the client side (#840)